refactor: update tabs block names and dependencies
