+++ packages/block-library/src/editor.scss
@@ -49,8 +49,8 @@
 @use "./social-links/editor.scss" as *;
 @use "./spacer/editor.scss" as *;
 @use "./table/editor.scss" as *;
-@use "./tabs-menu/editor.scss" as *;
-@use "./tabs-menu-item/editor.scss" as *;
+@use "./tab-list/editor.scss" as *;
+@use "./tab/editor.scss" as *;
 @use "./template-part/editor.scss" as *;
 @use "./term-template/editor.scss" as *;
 @use "./text-columns/editor.scss" as *;
