+++ packages/block-library/src/style.scss
@@ -76,10 +76,10 @@
 @use "./tag-cloud/style.scss" as *;
 @use "./tab/style.scss" as *;
 @use "./tab-panel/style.scss" as *;
+@use "./tab-panels/style.scss" as *;
 @use "./table/style.scss" as *;
 @use "./table-of-contents/style.scss" as *;
 @use "./tabs/style.scss" as *;
-@use "./tabs-menu-item/style.scss" as *;
 @use "./term-count/style.scss" as *;
 @use "./term-description/style.scss" as *;
 @use "./term-name/style.scss" as *;
