+++ packages/block-library/src/tab-list/block.json
@@ -2,13 +2,13 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"__experimental": true,
 	"apiVersion": 3,
-	"name": "core/tabs-menu",
+	"name": "core/tab-list",
 	"title": "Tab List",
 	"description": "Display the tab buttons for a tabbed interface.",
 	"category": "design",
 	"textdomain": "default",
 	"parent": [ "core/tabs" ],
-	"allowedBlocks": [ "core/tabs-menu-item" ],
+	"allowedBlocks": [ "core/tab" ],
 	"usesContext": [ "core/tabs-list" ],
 	"attributes": {},
 	"supports": {
