+++ packages/block-library/src/tab-list/editor.scss
@@ -1,10 +1,10 @@
-.wp-block-tabs-menu {
+.wp-block-tab-list {
 	// Allow the inner block list to be displayed in the flex layout, so the tab buttons appear as direct flex children in the editor.
 	> .block-editor-block-list__layout {
 		display: contents;
 	}
 
-	.block-editor-block-list__block:has(> .wp-block-tabs-menu-item) {
+	.block-editor-block-list__block:has(> .wp-block-tab) {
 		display: contents;
 	}
 }
