+++ packages/block-library/src/tab-panel/edit.js
@@ -1,42 +1,145 @@
 /**
  * WordPress dependencies
  */
+import { __ } from '@wordpress/i18n';
 import {
 	useBlockProps,
 	useInnerBlocksProps,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
-import { useSelect } from '@wordpress/data';
+import { useSelect, useDispatch } from '@wordpress/data';
+import { useMemo, useRef, useEffect } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import AddTabToolbarControl from '../tab/add-tab-toolbar-control';
-import RemoveTabToolbarControl from '../tab/remove-tab-toolbar-control';
+import Controls from './controls';
 
-const TAB_PANELS_TEMPLATE = [ [ 'core/tab', {} ] ];
+const TEMPLATE = [
+	[
+		'core/paragraph',
+		{
+			placeholder: __( 'Type / to choose a block' ),
+		},
+	],
+];
 
-export default function Edit( { clientId } ) {
-	const blockProps = useBlockProps();
+const { cancelAnimationFrame } = window;
 
-	const innerBlocksProps = useInnerBlocksProps( blockProps, {
-		template: TAB_PANELS_TEMPLATE,
-		templateLock: false,
-		renderAppender: false, // Appender handled by individual tab blocks
-	} );
+export default function Edit( { clientId, context, isSelected } ) {
+	const focusRef = useRef();
 
-	// Get the parent tabs block clientId
-	const tabsClientId = useSelect(
-		( select ) =>
-			select( blockEditorStore ).getBlockRootClientId( clientId ),
+	// Consume tab indices from context
+	const activeTabIndex = context[ 'core/tabs-activeTabIndex' ];
+	const editorActiveTabIndex = context[ 'core/tabs-editorActiveTabIndex' ];
+	const effectiveActiveIndex = editorActiveTabIndex ?? activeTabIndex;
+
+	// Clean up animation frames on unmount.
+	useEffect( () => {
+		return () => {
+			if ( focusRef.current ) {
+				cancelAnimationFrame( focusRef.current );
+			}
+		};
+	}, [] );
+
+	const { blockIndex, hasInnerBlocksSelected, tabsClientId } = useSelect(
+		( select ) => {
+			const {
+				getBlockRootClientId,
+				getBlockIndex,
+				hasSelectedInnerBlock,
+			} = select( blockEditorStore );
+
+			// Get the tab-panel parent first
+			const tabPanelClientId = getBlockRootClientId( clientId );
+			// Then get the tabs parent
+			const _tabsClientId = getBlockRootClientId( tabPanelClientId );
+
+			// Get data about this instance of core/tab.
+			const _blockIndex = getBlockIndex( clientId );
+			const _hasInnerBlocksSelected = hasSelectedInnerBlock(
+				clientId,
+				true
+			);
+
+			return {
+				blockIndex: _blockIndex,
+				hasInnerBlocksSelected: _hasInnerBlocksSelected,
+				tabsClientId: _tabsClientId,
+			};
+		},
 		[ clientId ]
 	);
 
+	const { updateBlockAttributes, __unstableMarkNextChangeAsNotPersistent } =
+		useDispatch( blockEditorStore );
+
+	// Sync editorActiveTabIndex when this tab is selected directly
+	useEffect( () => {
+		// Only update if this tab is selected and not already the active index
+		const isTabSelected = isSelected || hasInnerBlocksSelected;
+		if (
+			isTabSelected &&
+			tabsClientId &&
+			effectiveActiveIndex !== blockIndex
+		) {
+			// Mark as non-persistent so it doesn't add to undo history
+			__unstableMarkNextChangeAsNotPersistent();
+			updateBlockAttributes( tabsClientId, {
+				editorActiveTabIndex: blockIndex,
+			} );
+		}
+	}, [
+		isSelected,
+		hasInnerBlocksSelected,
+		tabsClientId,
+		effectiveActiveIndex,
+		blockIndex,
+		updateBlockAttributes,
+		__unstableMarkNextChangeAsNotPersistent,
+	] );
+
+	// Determine if this is the currently active tab (for editor visibility)
+	const isActiveTab = effectiveActiveIndex === blockIndex;
+
+	// Determine if this is the default tab (for the "Default Tab" toggle in controls)
+	const isDefaultTab = activeTabIndex === blockIndex;
+
+	/**
+	 * This hook determines if the current tab panel should be visible.
+	 * This is true if it is the editor active tab, or if it is selected directly.
+	 */
+	const isSelectedTab = useMemo( () => {
+		// Show if this tab is directly selected or has selected inner blocks
+		if ( isSelected || hasInnerBlocksSelected ) {
+			return true;
+		}
+		// Always show the active tab (at effectiveActiveIndex) regardless of other selection state.
+		// This ensures the tab panel remains visible when editing labels in tab-list.
+		if ( isActiveTab ) {
+			return true;
+		}
+		return false;
+	}, [ isSelected, hasInnerBlocksSelected, isActiveTab ] );
+
+	const blockProps = useBlockProps( {
+		hidden: ! isSelectedTab,
+		tabIndex: isSelectedTab ? 0 : -1,
+	} );
+
+	const innerBlocksProps = useInnerBlocksProps( blockProps, {
+		template: TEMPLATE,
+	} );
+
 	return (
-		<>
-			<AddTabToolbarControl tabsClientId={ tabsClientId } />
-			<RemoveTabToolbarControl tabsClientId={ tabsClientId } />
-			<div { ...innerBlocksProps } />
-		</>
+		<section { ...innerBlocksProps }>
+			<Controls
+				tabsClientId={ tabsClientId }
+				blockIndex={ blockIndex }
+				isDefaultTab={ isDefaultTab }
+			/>
+			{ isSelectedTab && innerBlocksProps.children }
+		</section>
 	);
 }
