+++ packages/block-library/src/tab-panels/edit.js
@@ -11,24 +11,26 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import AddTabToolbarControl from '../tab/add-tab-toolbar-control';
-import RemoveTabToolbarControl from '../tab/remove-tab-toolbar-control';
+import AddTabToolbarControl from '../tab-panel/add-tab-toolbar-control';
+import RemoveTabToolbarControl from '../tab-panel/remove-tab-toolbar-control';
 
-function Edit( { clientId } ) {
+const TAB_PANELS_TEMPLATE = [ [ 'core/tab-panel', {} ] ];
+
+export default function Edit( { clientId } ) {
+	const blockProps = useBlockProps();
+
+	const innerBlocksProps = useInnerBlocksProps( blockProps, {
+		template: TAB_PANELS_TEMPLATE,
+		templateLock: false,
+		renderAppender: false, // Appender handled by individual tab blocks
+	} );
+
+	// Get the parent tabs block clientId
 	const tabsClientId = useSelect(
 		( select ) =>
 			select( blockEditorStore ).getBlockRootClientId( clientId ),
 		[ clientId ]
 	);
-
-	const blockProps = useBlockProps();
-
-	const innerBlocksProps = useInnerBlocksProps( blockProps, {
-		allowedBlocks: [ 'core/tabs-menu-item' ],
-		orientation: 'horizontal',
-		templateLock: false,
-		renderAppender: false,
-	} );
 
 	return (
 		<>
@@ -38,5 +40,3 @@ function Edit( { clientId } ) {
 		</>
 	);
 }
-
-export default Edit;
