+++ packages/block-library/src/tab-panels/save.js
@@ -4,10 +4,8 @@
 import { useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
 
 export default function save() {
-	const blockProps = useBlockProps.save( {
-		role: 'tabpanel',
-	} );
+	const blockProps = useBlockProps.save();
 	const innerBlocksProps = useInnerBlocksProps.save( blockProps );
 
-	return <section { ...innerBlocksProps } />;
+	return <div { ...innerBlocksProps } />;
 }
