+++ packages/block-library/src/tab-panels/style.scss
@@ -1,0 +1,4 @@
+.wp-block-tab-panels {
+	flex-grow: 1;
+	min-width: 0;
+}
