+++ packages/block-library/src/tab/save.js
@@ -1,13 +1,13 @@
 /**
  * WordPress dependencies
  */
-import { useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
+import { useBlockProps } from '@wordpress/block-editor';
 
 export default function save() {
 	const blockProps = useBlockProps.save( {
-		role: 'tabpanel',
+		type: 'button',
+		role: 'tab',
 	} );
-	const innerBlocksProps = useInnerBlocksProps.save( blockProps );
 
-	return <section { ...innerBlocksProps } />;
+	return <button { ...blockProps } />;
 }
