--- packages/block-library/src/tabs-menu-item/controls.js
@@ -1,19 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { BlockControls } from '@wordpress/block-editor';
-
-/**
- * Internal dependencies
- */
-import AddTabToolbarControl from '../tab/add-tab-toolbar-control';
-import RemoveTabToolbarControl from '../tab/remove-tab-toolbar-control';
-
-export default function Controls( { tabsClientId } ) {
-	return (
-		<BlockControls>
-			<AddTabToolbarControl tabsClientId={ tabsClientId } />
-			<RemoveTabToolbarControl tabsClientId={ tabsClientId } />
-		</BlockControls>
-	);
-}
