--- packages/block-library/src/tabs-menu-item/style.scss
@@ -1,40 +1,0 @@
-.wp-block-tabs-menu-item {
-	box-sizing: border-box;
-	color: inherit;
-	display: block;
-	width: max-content;
-	text-decoration: none;
-	cursor: pointer;
-	flex-basis: inherit !important;
-	flex-grow: inherit !important;
-	position: relative;
-
-	// Button reset
-	border: none;
-	background: none;
-	appearance: none;
-	-webkit-appearance: none;
-
-	margin: 0;
-	padding: var(--wp--preset--spacing--20, 0.5em) var(--wp--preset--spacing--30, 1em);
-
-	// Inherit typography from parent
-	font-size: inherit;
-	font-family: inherit;
-	font-weight: inherit;
-	line-height: inherit;
-	letter-spacing: inherit;
-	text-transform: inherit;
-	text-align: inherit;
-
-	&[aria-selected="true"]::before,
-	&.is-active::before {
-		content: "";
-		position: absolute;
-		border-bottom: 2px solid currentColor;
-		pointer-events: none;
-		left: 0;
-		width: 100%;
-		bottom: 0;
-	}
-}
