+++ packages/block-library/src/tabs/block.json
@@ -7,7 +7,7 @@
 	"description": "Display content in a tabbed interface to help users navigate detailed content with ease.",
 	"category": "design",
 	"textdomain": "default",
-	"allowedBlocks": [ "core/tabs-menu", "core/tab-panel" ],
+	"allowedBlocks": [ "core/tab-list", "core/tab-panels" ],
 	"attributes": {
 		"activeTabIndex": {
 			"type": "number",
