+++ packages/icons/src/manifest.json
@@ -1584,19 +1584,19 @@
 		"filePath": "library/tab.svg"
 	},
 	{
+		"slug": "tab-list",
+		"label": "Tab List",
+		"filePath": "library/tab-list.svg"
+	},
+	{
+		"slug": "tab-panel",
+		"label": "Tab Panel",
+		"filePath": "library/tab-panel.svg"
+	},
+	{
 		"slug": "tabs",
 		"label": "Tabs",
 		"filePath": "library/tabs.svg"
-	},
-	{
-		"slug": "tabs-menu",
-		"label": "Tabs Menu",
-		"filePath": "library/tabs-menu.svg"
-	},
-	{
-		"slug": "tabs-menu-item",
-		"label": "Tabs Menu Item",
-		"filePath": "library/tabs-menu-item.svg"
 	},
 	{
 		"slug": "tag",
